refactor(ui): extract settings/data/config-loader.ts — load/render/collect/isProtectedKey
